moved a few more functions to utils
