+++ lib/board.mjs
@@ -1,20 +1,5 @@
 import { rowsFail, colsFail, housesFail } from './rules.mjs'
-
-function setCharAt(str, index, newChar) {
-  const before = str.substring(0, index)
-  const after = str.substring(index + 1)
-  return before + newChar + after
-}
-
-function xy2r(x, y) {
-  return x + y * 9
-}
-
-function r2xy(r) {
-  const x = r % 9
-  const y = Math.floor(r / 9)
-  return [x, y]
-}
+import { setCharAt, xy2r, r2xy } from './utils.mjs'
 
 export class Board {
   constructor(rows) {
