+++ lib/utils.mjs
@@ -4,6 +4,22 @@ export function times9(fn) {
   }
 }
 
+export function setCharAt(str, index, newChar) {
+  const before = str.substring(0, index)
+  const after = str.substring(index + 1)
+  return before + newChar + after
+}
+
+export function xy2r(x, y) {
+  return x + y * 9
+}
+
+export function r2xy(r) {
+  const x = r % 9
+  const y = Math.floor(r / 9)
+  return [x, y]
+}
+
 function getImportantCells(x, y) {
   let ret = [];
   function addIfNotXY(a, b) {
